third_party/blink/web_tests/svg/custom/use-cycle-with-attr-eventhandler-nested-expected.txt

CONSOLE MESSAGE: foo is called
CONSOLE MESSAGE: foo is called
